Send dice rolls to Discord
Send the dice roll from the character sheet to the Discord channel using the Webhook API. Using Webhooks is easier than using the full-fledged Bot and Gateway API for Discord.

Let the user specify which webhook URL to use when sending messages. Instead of hardcoding the URL in the codebase, which is both unsecure and limiting, the user is able to go into the extension's Options page and specify a URL themselves.