Report natural 20 roll
When a player rolls a 20 on an attack roll (without disadvantage) then send a message to the channel saying that the character rolled a natural 20.

According to the Player's Handbook for 5th Edition, [only attack rolls](https://rpg.stackexchange.com/a/108079) and death saving throws benefit from a natural 20. And when [rolling with disadvantage](https://rpg.stackexchange.com/a/72046), you must take the lower roll.

This does not take into account the scenario where the player rolls with disadvantage but rolls two 20s.