Add character image and roll info to message
Creates a Discord message with an embed object which will display the character's profile image, the dice roll description, breakdown, result, and notation.

Providing a profile image and roll description allows the results to be more readable when multiple players are rolling at once. Displaying the dice roll breakdown also provides more flexibility to the players and DM because it allows them to apply house rules to things like natural 20s and 1s.